Updates token data to save event instead of user_id
